# Improve the performance of certain GraphNode methods and avoid recalculating the AABB multiple times during rendering.