--- a/src/core/object-pool.js
+++ b/src/core/object-pool.js
@@ -30,6 +30,15 @@ class ObjectPool {
     _count = 0;
 
     /**
+     * A map from object references to their index in `_pool`. This is used to determine
+     * whether an object is actually allocated from this pool (and at which index).
+     *
+     * @type {WeakMap<InstanceType<T>, number>}
+     * @private
+     */
+    _objToIndexMap = new WeakMap();
+
+    /**
      * @param {T} constructorFunc - The constructor function for the
      * objects in the pool.
      * @param {number} size - The initial number of object instances to allocate.
@@ -41,18 +50,6 @@ class ObjectPool {
     }
 
     /**
-     * @param {number} size - The number of object instances to allocate.
-     * @private
-     */
-    _resize(size) {
-        if (size > this._pool.length) {
-            for (let i = this._pool.length; i < size; i++) {
-                this._pool[i] = new this._constructor();
-            }
-        }
-    }
-
-    /**
      * Returns an object instance from the pool. If no instances are available, the pool will be
      * doubled in size and a new instance will be returned.
      *
@@ -60,9 +57,41 @@ class ObjectPool {
      */
     allocate() {
         if (this._count >= this._pool.length) {
-            this._resize(this._pool.length * 2);
+            this._resize(Math.max(1, this._pool.length * 2));
         }
         return this._pool[this._count++];
+    }
+
+    /**
+     * Attempts to free the given object back into the pool. This only works if the object
+     * was previously allocated (i.e. it exists in `_objToIndexMap`) and is still in use
+     * (i.e., its index is below `_count`).
+     *
+     * @param {InstanceType<T>} obj
+     * @returns {boolean} Whether freeing succeeded.
+     */
+    free(obj) {
+        const index = this._objToIndexMap.get(obj);
+        if (index === undefined) {
+            return false;
+        }
+
+        if (index >= this._count) {
+            return false;
+        }
+
+        // Swap this object with the last allocated object, then decrement `_count`
+        const lastIndex = this._count - 1;
+        const lastObj = this._pool[lastIndex];
+
+        this._pool[index] = lastObj;
+        this._pool[lastIndex] = obj;
+
+        this._objToIndexMap.set(lastObj, index);
+        this._objToIndexMap.set(obj, lastIndex);
+
+        this._count -= 1;
+        return true;
     }
 
     /**
@@ -71,6 +100,21 @@ class ObjectPool {
      */
     freeAll() {
         this._count = 0;
+    }
+
+    /**
+     * @param {number} size - The number of object instances to allocate.
+     * @private
+     */
+    _resize(size) {
+        if (size > this._pool.length) {
+            for (let i = this._pool.length; i < size; i++) {
+                const obj = new this._constructor();
+                this._pool[i] = obj;
+
+                this._objToIndexMap.set(obj, i);
+            }
+        }
     }
 }