--- a/src/scene/graph-node.js
+++ b/src/scene/graph-node.js
@@ -75,7 +75,7 @@ function createTest(attr, value) {
 class GraphNode extends EventHandler {
     /**
      * @type {GraphNode[]}
-     * @ignore
+     * @private
      */
     static _stack = [];