… src/scene/mesh-instance.js
… src/scene/renderer/renderer.js
… src/scene/renderer/shadow-renderer.js
… test/core/object-pool.test.mjs
… test/core/queue.test.mjs