I experimented with placing multiple animated skinned meshes in Playcanvas and Unity 6.1, which is going to support WebGPU to address huge performance issues with skinned meshes. 

I understand that it's best to bake animations for skinned meshes instead of using the Anim component when dealing with high number of animated skinned meshes. But still... 

The PR proposes replacing the recursive implementation of core GraphNode methods with an iterative approach and avoid recalculating the AABB multiple times during rendering to improve performance of skinned meshes.

Here are my results:

| Main branch | PR |
|-------------|----|
| <img width="197" alt="Screenshot 2025-01-03 at 11 31 08" src="https://github.com/user-attachments/assets/f11f8822-ae10-4db3-be42-1e3fefbaa96f" /> | <img width="197" alt="Screenshot 2025-01-03 at 12 52 36" src="https://github.com/user-attachments/assets/ab543fcc-5d86-4699-b7d4-4345c1c911f4" /> |