@querielo 2025-01-03
@Maksims It is a good point, that's why _tmpGraphNodeStack was added. It doesn's looks like there is a big difference in terms of memory

| Main | PR |
| ----- | ---|
| <img width="837" alt="Screenshot 2025-01-03 at 14 39 48" src="https://github.com/user-attachments/assets/562c2082-1fd2-443f-bcc4-9ea1bfb4512f" /> | <img width="837" alt="Screenshot 2025-01-03 at 14 40 02" src="https://github.com/user-attachments/assets/f4632862-d264-4880-880f-c999562bc605" /> |