@willeastcott 2025-01-03
This is very cool. I was wondering...since this is such performance critical code, would it be more efficient to store the stack pointer (index)? Then, a pop is a decrement and a push is an array 'write at index'. Obviously, Array#push and Array#pop are very optimized...so I don't know if this would really make a tangible difference.