@AlexAPPi 2025-01-03
This approach eats up a lot of memory, maybe we should move away from storing references to objects in arrays on indexes?

https://www.mattzeunert.com/2018/01/25/v8-javascript-memory-quiz.html

```
interface Node {
    parent: Node;
    children: Node[];
}

// @ts-ignore
export function everyChildWithDive(entity: pc.Entity, callback: (child: pc.Entity, index: number, depth: number) => void | 1 | 2, offset?: number): void | 2;
export function everyChildWithDive<T extends Node>(node: T, callback: (child: T, index: number, depth: number) => void | 1 | 2, offset?: number): void | 2;
export function everyChildWithDive(node: Node, callback: (child: Node, index: number, depth: number) => void | 1 | 2, offset: number = 0): void | 2 {

    let children = node.children;
    let length   = children.length;

    if (length < 1) {
        return;
    }

    let parent = node;
    let stackIndex = 0;

    const stack = [offset];

    while (offset < length) {

        const child  = children[offset];
        const result = callback(child, offset, stackIndex);

        if (result === 2) {
            return 2;
        }
        
        stack[stackIndex] = ++offset;

        if (result !== 1) {

            const childChildren = child.children;
            const childChildrenLength = childChildren.length;

            if (childChildrenLength > 0) {

                offset   = 0;
                parent   = child;
                children = childChildren;
                length   = childChildrenLength;

                stackIndex++;
                stack.push(offset);

                continue;
            }
        }

        while (offset === length && stackIndex > 0) {
            stack.pop();
            offset   = stack[--stackIndex];
            parent   = parent.parent;
            children = parent.children;
            length   = children.length;
        }
    }
}
```