@querielo 2025-01-03
@AlexAPPi  I'll review your suggestion tomorrow, ok?

Regarding my solution, as noted in the comment, I'm using shift() to implement a breadth-first approach (it is a queue instead of a stack):

```javascript
const queue = [node];
while (queue.length > 0) {
    // shift() is used for breadth-first approach; use pop() for depth-first
    const current = queue.shift();
    
    // . . .
    
    const children = current.children;
    for (let i = 0; i < children.length; i++) {
        queue.push(children[i]);
    }
}
```

While **breadth-first traversal** tends to be slower and consume more memory than **depth-first**, it's necessary here to pass unit tests that expect a specific order. I'd prefer to use `pop()` for a **depth-first approach**, but this would require modifying some unit tests. I'm hesitant to make such changes, as they might disrupt components that depend on the current order.

Note: It appears that the third solution from https://jsbench.me/cvm5h498pk/1 has a bug (I'm not sure about purpose of the `break` statement).

<img width="986" alt="Screenshot 2025-01-04 at 00 04 56" src="https://github.com/user-attachments/assets/f06c1308-c922-4615-91b3-816ef967fb67" />

I cannot run the fourth solution:

<img width="986" alt="Screenshot 2025-01-04 at 00 07 57" src="https://github.com/user-attachments/assets/5710a270-cd61-4830-8068-aa0d42f7642b" />

But as I can see the **depth-first** solution is faster than the second solution:

<img width="825" alt="Screenshot 2025-01-04 at 00 10 41" src="https://github.com/user-attachments/assets/3ff6b99a-8eef-4935-a585-75a42bc55e45" />

Thank you for [this article](https://www.mattzeunert.com/2018/01/25/v8-javascript-memory-quiz.html)!

@willeastcott  can I change **breadth-first** to **depth-first**? And a unit test that requires **breadth-first**?