@querielo 2025-01-04
<img width="988" alt="Screenshot 2025-01-04 at 13 04 25" src="https://github.com/user-attachments/assets/338245dd-fd13-4f19-a3b7-176532ad07ba" />

I'll try to implement a solution with a queue pool (breadth-first). It is 20% slower than using a pure offset, but requires less memory. For example, on my scene with 21_943 nodes a capacity of an auto extending circular queue is 4098. 

<img width="1065" alt="Screenshot 2025-01-04 at 13 18 58" src="https://github.com/user-attachments/assets/e283aaab-2a84-4cce-8033-b32e97b279ae" />