@querielo 2025-01-05
Update: 
- Introduce a new `Queue` class that implements a circular buffer, automatically resizing its capacity as needed.
  - Added tests for `Queue`
- Add a `free` method to `ObjectPool` to allow deallocation of objects back into the pool.
  - Refactor internal `_resize` logic in `ObjectPool` to track object indices, enabling safer and more efficient memory management.
  - Added tests for `ObjectPool`
- Update `GraphNode.forEach` to use a pooled `Queue` for iteration, reducing memory allocations in breadth-first traversals.