@Maksims 2025-01-06
Changing the order of execution in public APIs definitely will break people's project, so we always avoid any breaking changes of public APIs.