@querielo 2025-01-18
@willeastcott  Unfortunately, I had not enough time. Still WIP

I make experiments with [this benchmark](https://gist.github.com/querielo/80fb69755e0c94b9bc0d4301e0d0aea7#file-gistfile1-txt). My best result is here:

| Test Case                         |  Main (ops/sec ± %)               | Research result WIP (ops/sec ± %)              |
|-----------------------------------|------------------------------------------|------------------------------------------|
| complex-16: findByName (p0)       | 41,631,363 ±17.40% (84 runs sampled)    | 82,724,354 ±10.49% (85 runs sampled)    |
| complex-16: findByName (p5)       | 17,509 ±0.45% (98 runs sampled)         | 38,081 ±1.38% (92 runs sampled)         |
| complex-16: findByName (p10)      | 8,430 ±1.46% (88 runs sampled)          | 21,226 ±0.93% (92 runs sampled)         |
| complex-16: findByName (p25)      | 1,865 ±1.77% (89 runs sampled)          | 8,564 ±0.72% (96 runs sampled)          |
| complex-16: findByName (p50)      | 607 ±1.43% (90 runs sampled)            | 1,787 ±3.40% (89 runs sampled)          |
| complex-16: findByName (p75)      | 387 ±1.52% (90 runs sampled)            | 787 ±1.87% (88 runs sampled)            |
| complex-16: findByName (p90)      | 276 ±2.44% (83 runs sampled)            | 636 ±2.01% (93 runs sampled)            |
| complex-16: findByName (p95)      | 230 ±8.46% (78 runs sampled)            | 685 ±1.17% (95 runs sampled)            |
| complex-16: findByName (p100)     | 225 ±1.55% (79 runs sampled)            | 568 ±1.32% (94 runs sampled)            |
| complex-16: forEach - count "child" | 422 ±1.91% (90 runs sampled)            | 3,062 ±1.37% (98 runs sampled)          |
| deep-narrow: findByName (p0)      | 51,111,560 ±0.59% (100 runs sampled)    | 75,038,398 ±0.72% (99 runs sampled)     |
| deep-narrow: findByName (p5)      | 15,686,485 ±1.24% (99 runs sampled)     | 17,354,031 ±0.55% (96 runs sampled)     |
| deep-narrow: findByName (p10)     | 7,481,730 ±0.77% (99 runs sampled)      | 7,871,519 ±0.34% (99 runs sampled)      |
| deep-narrow: findByName (p25)     | 4,719,099 ±0.30% (98 runs sampled)      | 5,284,981 ±0.44% (98 runs sampled)      |
| deep-narrow: findByName (p50)     | 1,830,604 ±0.29% (97 runs sampled)      | 1,719,027 ±4.15% (94 runs sampled)      |
| deep-narrow: findByName (p75)     | 1,237,125 ±0.41% (95 runs sampled)      | 1,259,369 ±0.87% (93 runs sampled)      |
| deep-narrow: findByName (p90)     | 987,077 ±0.80% (100 runs sampled)       | 1,003,850 ±0.39% (98 runs sampled)      |
| deep-narrow: findByName (p95)     | 879,433 ±4.00% (98 runs sampled)        | 925,820 ±0.84% (95 runs sampled)        |
| deep-narrow: findByName (p100)    | 828,371 ±3.59% (95 runs sampled)        | 869,562 ±0.43% (98 runs sampled)        |
| deep-narrow: forEach - count "child" | 1,410,805 ±0.47% (97 runs sampled)      | 1,260,831 ±0.95% (97 runs sampled)      |
| wide-shallow: findByName (p0)     | 44,312,496 ±3.26% (93 runs sampled)     | 68,127,657 ±1.31% (89 runs sampled)     |
| wide-shallow: findByName (p5)     | 11,677,169 ±0.63% (96 runs sampled)     | 11,708,320 ±0.54% (95 runs sampled)     |
| wide-shallow: findByName (p10)    | 7,529,309 ±0.34% (98 runs sampled)      | 8,772,932 ±0.35% (101 runs sampled)     |
| wide-shallow: findByName (p25)    | 2,439,713 ±0.28% (100 runs sampled)     | 2,325,104 ±0.96% (96 runs sampled)      |
| wide-shallow: findByName (p50)    | 1,198,470 ±0.42% (96 runs sampled)      | 1,059,050 ±0.31% (100 runs sampled)     |
| wide-shallow: findByName (p75)    | 782,697 ±0.23% (99 runs sampled)        | 703,360 ±2.61% (96 runs sampled)        |
| wide-shallow: findByName (p90)    | 856,695 ±0.80% (97 runs sampled)        | 936,670 ±0.80% (93 runs sampled)        |
| wide-shallow: findByName (p95)    | 816,399 ±0.44% (93 runs sampled)        | 849,755 ±0.31% (101 runs sampled)       |
| wide-shallow: findByName (p100)   | 750,973 ±0.47% (99 runs sampled)        | 732,667 ±2.51% (94 runs sampled)        |
| wide-shallow: forEach - count "child" | 868,954 ±2.90% (94 runs sampled)        | 817,848 ±0.74% (98 runs sampled)        |
| ultra-deep: findByName (p0)       | 44,595,915 ±1.08% (96 runs sampled)     | 84,905,349 ±1.67% (95 runs sampled)     |
| ultra-deep: findByName (p5)       | 148,084 ±0.30% (98 runs sampled)        | 150,822 ±0.67% (97 runs sampled)        |
| ultra-deep: findByName (p10)      | 58,782 ±0.40% (96 runs sampled)         | 64,269 ±0.22% (100 runs sampled)        |
| ultra-deep: findByName (p25)      | 30,092 ±0.40% (96 runs sampled)         | 32,706 ±0.40% (99 runs sampled)         |
| ultra-deep: findByName (p50)      | 17,164 ±0.24% (100 runs sampled)        | 17,338 ±0.41% (99 runs sampled)         |
| ultra-deep: findByName (p75)      | 11,885 ±0.56% (99 runs sampled)         | 11,500 ±4.24% (95 runs sampled)         |
| ultra-deep: findByName (p90)      | 10,111 ±0.51% (99 runs sampled)         | 9,268 ±4.93% (94 runs sampled)          |
| ultra-deep: findByName (p95)      | 12,231 ±0.47% (100 runs sampled)        | 12,603 ±2.66% (85 runs sampled)         |
| ultra-deep: findByName (p100)     | 13,294 ±1.06% (98 runs sampled)         | 14,171 ±3.56% (94 runs sampled)         |
| ultra-deep: forEach - count "child" | 15,832 ±0.54% (100 runs sampled)        | 14,877 ±2.46% (99 runs sampled)         |

The test cases measure the performance of different hierarchical structures and operations on them.

* complex-16: A mixed-depth structure with levels of varying branching factors. Simulates realistic complex structures, such as file systems or dependency trees.
* deep-narrow: A tall and thin structure, where depth is emphasized. Useful for testing operations where recursion or depth-first search could be expensive.
* wide-shallow: A flat and wide structure, emphasizing breadth. Tests operations that deal with many siblings in a single level.
* ultra-deep: An extremely deep structure, testing the limits of recursion or stack depth in operations.

p* represents the node at *% of the total hierarchy traversal order, calculated after converting the tree into a linear array.
For example, p10, p25, represent nodes at 10%, 25%, etc., of the array, providing a way to test performance across different levels or positions in the tree.

Result: my WIP code solves nearly all of @Maksims' concerns. However, the current implementation is somewhat unrefined and requires further improvement before publishing. Also, it's essential to conduct additional testing to enhance code quality and reliability and expand the suite of unit tests.

I added [this PR](https://github.com/playcanvas/engine/pull/7296) about AABB.